requirements.txt
CoffeeScript==1.0.5
Mako==0.7.3
MarkupSafe==0.15
Paste==1.7.5.1
Plim==0.7.11
PyExecJS==1.0.4
WebOb==1.2.3
argparse==1.2.1
distribute==0.6.24
docutils==0.10
facebook-sdk==0.4.0
html5lib==0.95
markdown2==2.1.0
pyScss==1.1.5
pymongo==2.5
stylus==0.1.0
webapp2==2.5.2
wsgiref==0.1.2
